+++ src/aim/models/__init__.py
@@ -15,9 +15,12 @@ def deepcopy_except_parent(obj):
     """Returns a deepcopy on on the object, but does not recurse up the
     __parent__ attribute to prevent copying the entire model.
     """
-    parent = obj.__parent__
+    if not hasattr(obj, '__parent__'):
+        return copy.deepcopy(obj)
+
+    # set __parent__ to None, copy the object, then restore __parent__
+    parent = getattr(obj, '__parent__', None)
     obj.__parent__ = None
     copy_obj = copy.deepcopy(obj)
     obj.__parent__ = parent
     return copy_obj
-